+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Generation of fields for CSV source, if not given.
+- Removed unused code
 
 ## [1.0.0] - 2025-09-03